+++ src/storage.rs
@@ -24,7 +24,6 @@ use std::{path::PathBuf, sync::Arc};
 use thiserror::Error;
 
 use crate::{
-    array::ChunkKeyEncoding,
     byte_range::{ByteOffset, ByteRange, InvalidByteRangeError},
     node::{NodeNameError, NodePath, NodePathError},
 };
@@ -34,19 +33,16 @@ pub use store_prefix::{StorePrefix, StorePrefixError, StorePrefixes};
 
 #[cfg(feature = "async")]
 pub use self::storage_async::{
-    async_create_array, async_create_group, async_discover_children, async_discover_nodes,
-    async_erase_chunk, async_erase_metadata, async_erase_node, async_get_child_nodes,
-    async_node_exists, async_node_exists_listable, async_retrieve_chunk,
-    async_retrieve_partial_values, async_store_chunk, async_store_set_partial_values,
-    AsyncListableStorageTraits, AsyncReadableListableStorageTraits, AsyncReadableStorageTraits,
-    AsyncReadableWritableListableStorageTraits, AsyncReadableWritableStorageTraits,
-    AsyncWritableStorageTraits,
+    async_discover_children, async_discover_nodes, async_erase_node, async_get_child_nodes,
+    async_node_exists, async_node_exists_listable, async_retrieve_partial_values,
+    async_store_set_partial_values, AsyncListableStorageTraits, AsyncReadableListableStorageTraits,
+    AsyncReadableStorageTraits, AsyncReadableWritableListableStorageTraits,
+    AsyncReadableWritableStorageTraits, AsyncWritableStorageTraits,
 };
 
 pub use self::storage_sync::{
-    create_array, create_group, discover_children, discover_nodes, erase_chunk, erase_metadata,
-    erase_node, get_child_nodes, node_exists, node_exists_listable, retrieve_chunk,
-    retrieve_partial_values, store_chunk, store_set_partial_values, ListableStorageTraits,
+    discover_children, discover_nodes, erase_node, get_child_nodes, node_exists,
+    node_exists_listable, retrieve_partial_values, store_set_partial_values, ListableStorageTraits,
     ReadableListableStorageTraits, ReadableStorageTraits, ReadableWritableListableStorageTraits,
     ReadableWritableStorageTraits, WritableStorageTraits,
 };
@@ -285,8 +281,15 @@ fn meta_key_any(path: &NodePath, metadata_file_name: &str) -> StoreKey {
 }
 
 /// Return the Zarr V3 metadata key (zarr.json) given a node path.
+// TODO: Deprecate this
 #[must_use]
 pub fn meta_key(path: &NodePath) -> StoreKey {
+    meta_key_v3(path)
+}
+
+/// Return the Zarr V3 metadata key (zarr.json) given a node path.
+#[must_use]
+pub fn meta_key_v3(path: &NodePath) -> StoreKey {
     meta_key_any(path, "zarr.json")
 }
 
@@ -308,17 +311,15 @@ pub fn meta_key_v2_attributes(path: &NodePath) -> StoreKey {
     meta_key_any(path, ".zattrs")
 }
 
-/// Return the data key given a node path, chunk grid coordinates, and a chunk key encoding.
+/// Return the data key given a node path and a `chunk_key` of an array.
+///
+/// A chunk key is computed with the `encode` method of a chunk key encoder.
 #[must_use]
-pub fn data_key(
-    path: &NodePath,
-    chunk_grid_indices: &[u64],
-    chunk_key_encoding: &ChunkKeyEncoding,
-) -> StoreKey {
+pub fn data_key(path: &NodePath, chunk_key: &StoreKey) -> StoreKey {
     let path = path.as_str();
     let path = path.strip_prefix('/').unwrap_or(path);
     let mut key_path = PathBuf::from(path);
-    key_path.push(chunk_key_encoding.encode(chunk_grid_indices).as_str());
+    key_path.push(chunk_key.as_str());
     unsafe { StoreKey::new_unchecked(key_path.to_string_lossy().to_string()) }
 }