+++ CHANGELOG.md
@@ -33,6 +33,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
  - **Breaking**: Remove `array::NonZeroError`, use `std::num::TryFromIntError` instead
  - **Breaking**: Move storage transformers from the `storage` to the `array` namespace
+ - **Breaking**: Remove many functions in the storage namespace:
+    - `[async_]create_{array, group}`
+    - `[async_]erase_{chunk,metadata}`, `[async_]{retrieve,store}_chunk`
 
 ### Fixed
  - `[async_]store_set_partial_values` no longer truncates